core: MultiPointWorker: separate into its own file
